requirements.txt
asgiref==3.8.1
click==8.1.7
diff-match-patch==20241021
Django==5.1.3
django-debug-toolbar==4.4.6
django-import-export==4.3.1
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
ffmpeg==1.4
python-decouple==3.8
redis==5.2.0
rq==1.16.2
sqlparse==0.5.2
tablib==3.7.0
